lectures/2/requirements.txt
fastapi
uvicorn
requests
python-multipart
websockets
openai
python-dotenv
pydantic[email]